+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v0.2.0 (2022-11-22)
 * Added Controller structure
 * Updated Genesis for Fluidity Cast